# perf(connector): do not create span for parsing each row 