Signed-off-by: Bugen Zhao <i@bugenzhao.com>I hereby agree to the terms of the [RisingWave Labs, Inc. Contributor License Agreement](https://gist.github.com/TennyZhuang/f00be7f16996ea48effb049aa7be4d66#file-rw_cla).

## What's changed and what's your intention?

In #13073 and https://github.com/risingwavelabs/risingwave/issues/12959#issuecomment-1780750295, we find the performance overhead of constructing row-level tracing span is not acceptable. So we remove that in this PR, and manually pass the row context.

----

Benchmark shows that there's no cost for happy path, which is expected.

```
parse_nexmark           time:   [1.2238 ms 1.2273 ms 1.2309 ms]

parse_nexmark_with_tracing
                        time:   [1.2131 ms 1.2166 ms 1.2200 ms]
```

fix #12959

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Documentation

- [ ] My PR needs documentation updates. (Please use the **Release note** section below to summarize the impact on users)

## Release note

If this PR includes changes that directly affect users or other significant modifications relevant to the community, kindly draft a release note to provide a concise summary of these changes. Please prioritize highlighting the impact these changes will have on users.


<!--
Please create a release note for your changes.

Discuss technical details in the "What's changed" section, and
focus on the impact on users in the release note.

You should also mention the environment or conditions where the impact may occur.
-->

</details>